Moving shared memory data transfer related changes into separate directory
