+++ azure_functions_worker/shared_memory_data_transfer/file_accessor.py
