+++ azure_functions_worker/shared_memory_data_transfer/shared_memory_map.py
@@ -8,6 +8,7 @@ import struct
 import sys
 from typing import Optional
 from .shared_memory_constants import SharedMemoryConstants as consts
+from .file_accessor import FileAccessor
 from ..logging import logger
