+++ azure_functions_worker/shared_memory_data_transfer/shared_memory_metadata.py
